Create a DialogWrapper component to minimize  boilerplate when creating modals
